Fix 16-bit channels
Scenario: "Pan" may be channel 1 and "Pan fine" may be channel 3. The former system of marking 16-bit
channels as two-value arrays and additionally providing "MSB"/"LSB" information can not cover this case.
Additionally, it was not really nice to enter all that information.

The new system uses a new field in the fixture definition. See [fixtures_default.js](fixtures_default.js) and
look for `multiByteChannels` for more details.

Also, `heads` was moved from modes to fixture, as this is actually really close to `multiByteChannels` and
makes more sense there.